+++ CHANGELOG.md
@@ -18,6 +18,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
   its own envelope. See [`UPGRADE-5.0.md`](/UPGRADE-5.0.md) for more details.
 - `PMG\Queue\MessageHandler::handle` now typehints against `object` instead of
   `Message`.
+- `PMG\Queue\Producer::send` now typehints against `object` instead of `Message`.
 - `MessageLifecycle::failed` no longer has an `$isRetrying` argument, instead
   `MessageLifecycyle::retrying` will be called instead.
 - The `Message` interface no longer has a `getName` method.